refactor (attendance): move factory to its own file; move enumeration to rules.js;
