--- src/lib/attendance/enums.js
@@ -1,7 +1,0 @@
-export const RULE = {
-  EXCEPTION_MO: 'exception_mo',
-  DAYOFF: 'dayoff',
-  ONTIME: 'ontime',
-  WARNING: 'warning',
-  DANGER: 'danger'
-}
