+++ src/lib/attendance/index.js
@@ -1,2 +1,2 @@
-export * from './enums'
 export * from './rules'
+export * from './rules.factory'
