+++ src/lib/attendance/rules.js
@@ -1,84 +1,38 @@
-import _inRange from 'lodash/inRange'
-import addMinutes from 'date-fns/addMinutes'
-import { RULE } from './enums'
 import { ATTENDANCE } from '../constants'
+import { Rule, createTimeRangeRule } from './rules.factory'
 
-/**
- * Rule handler.
- * @callback RuleHandler
- * @param {Object} checkinData - array element of "/attendance" XHR response data.
- * @returns {boolean}
- */
-
-/**
- * Rule definition.
- * @typedef {Object} Rule
- * @property {string} [name]
- * @property {RuleHandler} handler
- */
-
-/**
- * Time range rule factory.
- * @param {string} start - HH:MM format
- * @param {string} end - HH:MM format
- * @returns {Rule}
- */
-function createTimeRangeRule (name, start, end) {
-  return {
-    name,
-    handler (checkinData) {
-      const { startDate } = checkinData
-      const value = new Date(startDate).setSeconds(0, 0)
-      const rangeStart = new Date(startDate).setHours(...start.split(':'), 0, 0)
-      const rangeEnd = new Date(startDate).setHours(...end.split(':'), 0, 0)
-
-      return _inRange(
-        +value,
-        addMinutes(+rangeStart, 1),
-        addMinutes(+rangeEnd, 1)
-      )
-    }
-  }
+export const RULE = {
+  DAYOFF: 'dayoff',
+  ONTIME: 'ontime',
+  WARNING: 'warning',
+  DANGER: 'danger'
 }
 
+const dayoff = new Rule(RULE.DAYOFF, (data) => {
+  return data.message !== ATTENDANCE.PRESENT
+})
+
 /**
- * If rule is not a type of time range rule,
- * rules must be ordered based on priority,
- * since rule is matched using match-first strategy.
- *
- * For example, dayoff should come before exception for MO.
  * @type {Rule[]}
  */
 const rules = [
-  {
-    name: RULE.DAYOFF,
-    handler (checkinData) {
-      return checkinData.message !== ATTENDANCE.PRESENT
-    }
-  },
-  {
-    name: RULE.EXCEPTION_MO,
-    handler (checkinData) {
-      // TODO: shouldn't be determined by string matching
-      return checkinData.jabatan === 'Monitoring Officer'
-    }
-  },
+  dayoff,
   createTimeRangeRule(RULE.ONTIME, '00:00', '07:30'),
   createTimeRangeRule(RULE.WARNING, '07:30', '09:00'),
   createTimeRangeRule(RULE.DANGER, '09:00', '24:00')
 ]
 
 /**
- * @param {Date} checkinDate
- * @returns {string | undefined} matched rule name
+ * @param {Object} checkinData - array element of "/attendance" XHR response data.
+ * @returns {string[]} matched rule names
  */
-export function validate (checkinDate) {
-  let result
-  for (const rule of rules) {
-    if (rule.handler(checkinDate)) {
-      result = rule.name
+export function getMatchedRule (checkinData) {
+  let rule
+  for (const r of rules) {
+    if (r.handler(checkinData)) {
+      rule = r.name
       break
     }
   }
-  return result
+  return rule
 }
